backend/requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
django-filter==21.1
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
djangorestframework-simplejwt==4.7.2
webcolors==1.13
Pillow==10.1.0